Rename erica request files
