+++ erica/domain/Shared/BaseDomainModel.py
@@ -3,7 +3,6 @@ from abc import ABCMeta
 from typing import TypeVar, Generic, Optional
 
 from pydantic import BaseModel
-from pydantic.main import BaseModel
 
 DataT = TypeVar('DataT')
 ClassT = TypeVar('ClassT')
